# src/api/swagger/swaggerJson.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the Swagger Documentation implementation of API's on NodeJS. You will find detailed documentation and model data about the REST API's. To check if an API is working , execute the API after entering the valid parameters",
		"version": "1.0.0",
		"title": "Boiler",
		"termsOfService": "",
		"contact": {
			"email": "api@boiler.com"
		}
	},
	"host": "127.0.0.1:4001",
	"basePath": "/api/v1",
...
	"schemes": [
		"http"
	],
	"paths": {
...
		"/admin/user/{userId}": {
			"get": {
				"tags": ["Admin"],
				"summary": "get an user info",
				"description": "get user information",
				"operationId": "getUser",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "userId",
						"description": "agent id",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully fetch user info"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"delete": {
				"tags": ["Admin"],
				"summary": "delete an user info",
				"description": "delete user information",
				"operationId": "getUser",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "userId",
						"description": "agent id",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully delete user info"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/user-list": {
			"post": {
				"tags": ["Admin"],
				"summary": "get user list",
				"description": "get user list",
				"operationId": "getUserList",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "register data to be added to db",
						"required": true,
						"schema": {
							"$ref": "#/definitions/userList"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successfully fetch user list"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/agent": {
			"post": {
				"tags": ["Admin"],
				"summary": "Register a new agent",
				"description": "Adds a new agent data information into the database",
				"operationId": "createAgent",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "register data to be added to db",
						"required": true,
						"schema": {
							"$ref": "#/definitions/createAgent"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully Created"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/agent/{agentId}": {
			"put": {
				"tags": ["Admin"],
				"summary": "Update an agent",
				"description": "update agent information",
				"operationId": "updateAgent",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "register data to be added to db",
						"required": true,
						"schema": {
							"$ref": "#/definitions/updateAgent"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
					,
					{
						"in": "path",
						"name": "agentId",
						"description": "agent id",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully Created"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/property-list/{page}/{limit}": {
			"get": {
				"tags": ["Admin"],
				"summary": "List of properties",
				"description": "List of properties",
				"operationId": "getPropertyList",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "page",
						"description": "page number",
						"required": true,
						"type": "integer"
					},
					{
						"in": "path",
						"name": "limit",
						"description": "limit of the page",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully getting list"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/property/{propertyId}": {
			"get": {
				"tags": ["Admin"],
				"summary": "Property details",
				"description": "get details of a property",
				"operationId": "getPropertyDetails",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "propertyId",
						"description": "Id of the property",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully getting property details"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/deal-list/{page}/{limit}": {
			"get": {
				"tags": ["Admin"],
				"summary": "List of deals",
				"description": "List of deals",
				"operationId": "getDealList",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "page",
						"description": "page number",
						"required": true,
						"type": "integer"
					},
					{
						"in": "path",
						"name": "limit",
						"description": "limit of the page",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully getting list"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/deal-details/{dealId}": {
			"get": {
				"tags": ["Admin"],
				"summary": "deal details",
				"description": "get details of a deal",
				"operationId": "dealDetails",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "dealId",
						"description": "Id of the deal",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully getting deal details"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/admin/assign-property-to-agent": {
			"post": {
				"tags": ["Admin"],
				"summary": "Assign a property to an agent",
				"description": "Assign a property to an agent",
				"operationId": "assignPropertyToAgent",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "register data to be added to db",
						"required": true,
						"schema": {
							"$ref": "#/definitions/assignPropertyToAgent"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully assigned"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
		"/seller/property/{id}": {
			"get": {
				"tags": [
					"SELLER"
				],
				"summary": " Fetch Specific Property",
				"description": "Fetch Specific Property",
				"operationId": "getProperty",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"delete": {
				"tags": [
					"SELLER"
				],
				"summary": " Soft delete specific property",
				"description": "Soft delete specific property",
				"operationId": "deleteProperty",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"patch": {
				"tags": [
					"SELLER"
				],
				"summary": " ico details by amdin",
				"description": "ico details list by amdin based on icoid",
				"operationId": "getIcoDetails via icoID",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Seller updates a property",
						"required": true,
						"schema": {
							"$ref": "#/definitions/updateProperty"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"userList": {
			"type": "object",
			"required": [
				"type"
			],
			"properties": {
				"type": {
					"type": "string",
					"enum": ["AGENT", "USER"]
				},
				"page": {
					"type": "integer",
					"example": 0
				},
				"limit": {
					"type": "integer",
					"example": 10
				}
			},
			"xml": {
				"name": "userList"
			}
		},
		"createAgent": {
			"type": "object",
			"required": [
				"firstName",
				"lastName",
				"email"
			],
			"properties": {
				"firstName": {
					"type": "string",
					"example": "john22"
				},
				"lastName": {
					"type": "string",
					"example": "john22"
				},
				"email": {
					"type": "string",
					"example": "john@blockgemini.com"
				},
				"phoneNumber": {
					"type": "number",
					"example": "7002708122"
				},
				"countryCode": {
					"type": "number",
					"example": "91"
				}
			},
			"xml": {
				"name": "createAgent"
			}
		},
		"updateAgent": {
			"type": "object",
			"required": [
				"firstName",
				"lastName",
				"email"
			],
			"properties": {
				"firstName": {
					"type": "string",
					"example": "john22"
				},
				"lastName": {
					"type": "string",
					"example": "john22"
				},
				"email": {
					"type": "string",
					"example": "john@blockgemini.com"
				},
				"phoneNumber": {
					"type": "number",
					"example": "7002708122"
				},
				"countryCode": {
					"type": "number",
					"example": "91"
				}
			},
			"xml": {
				"name": "updateAgent"
			}
		},
		"assignPropertyToAgent": {
			"type": "object",
			"required": [
				"agentId",
				"propertyId"
			],
			"properties": {
				"agentId": {
					"type": "string",
					"example": "sdfsdfsdf"
				},
				"propertyId": {
					"type": "string",
					"example": "jsdfgsgfskgds"
				}
			},
			"xml": {
				"name": "updateAgent"
			}
		},
...
		"updateProperty": {
			"type": "object",
			"properties": {
				"city": {
					"type": "string",
					"example": "ABC"
				},
				"locality": {
					"type": "string",
					"example": "DEF"
				},
				"address": {
					"type": "string",
					"example": "GHI"
				},
				"numberOfBathrooms": {
					"type": "number",
					"example": "1"
				},
				"numberOfBedrooms": {
					"type": "number",
					"example": "1"
				},
				"numberOfRooms": {
					"type": "number",
					"example": "1"
				},
				"livingArea": {
					"type": "number",
					"example": "2500"
				},
				"email": {
					"type": "string",
					"example": "john@snow.com"
				}
			},
			"xml": {
				"name": "updateProperty"
			}
		}
	}
}